Move character and search functions to their own objects
