+++ src/test/kotlin/profile/MinionsScrapeTest.kt
@@ -1,7 +1,7 @@
 package profile
 
 import cloud.drakon.ktlodestone.exception.CharacterNotFoundException
-import cloud.drakon.ktlodestone.getMinions
+import cloud.drakon.ktlodestone.Character.getMinions
 import kotlin.test.Test
 import kotlinx.coroutines.runBlocking
 import org.junit.jupiter.api.Assertions
